+++ tests/bitserializer_pugixml_tests/pugixml_archive_tests.cpp
@@ -46,7 +46,7 @@ TEST(PugiXmlArchive, SerializeTwoDimensionalArray) {
 
 TEST(PugiXmlArchive, ShouldLoadToArrayWithLesserAmountOfElements) {
 	TestSerializeArray<XmlArchive, bool, 7, 5>();
-	TestSerializeArray<XmlArchive, int64_t, 7, 5>();
+	TestSerializeArray<XmlArchive, int, 7, 5>();
 	TestSerializeArray<XmlArchive, double, 7, 5>();
 	TestSerializeArray<XmlArchive, std::string, 7, 5>();
 	TestSerializeArray<XmlArchive, TestPointClass, 7, 5>();
@@ -54,7 +54,7 @@ TEST(PugiXmlArchive, ShouldLoadToArrayWithLesserAmountOfElements) {
 
 TEST(PugiXmlArchive, ShouldLoadToArrayWithBiggerAmountOfElements) {
 	TestSerializeArray<XmlArchive, bool, 5, 7>();
-	TestSerializeArray<XmlArchive, int64_t, 5, 7>();
+	TestSerializeArray<XmlArchive, int, 5, 7>();
 	TestSerializeArray<XmlArchive, double, 5, 7>();
 	TestSerializeArray<XmlArchive, std::string, 5, 7>();
 	TestSerializeArray<XmlArchive, TestPointClass, 5, 7>();
